refactor(frontend): split course detail panel sections
